git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@14414 27425a3e-05d8-49a3-a47f-9c15f0e5edd8

+++ os/hal/ports/STM32/STM32L4xx+/hal_lld.c
@@ -371,7 +371,7 @@ static bool hal_lld_clock_check_tree(const halclkcfg_t *ccp) {
   }
 
   /* MSI clock.*/
-  msiidx = (uint8_t)((ccp->rcc_cr & RCC_CR_MSIRANGE) >> RCC_CR_MSIRANGE_Pos);
+  msiidx = (ccp->rcc_cr & RCC_CR_MSIRANGE) >> RCC_CR_MSIRANGE_Pos;
   if (msiidx >= STM32_MSIRANGE_ARRAY_SIZE) {
     return true;
   }
